be explicit for msvc
(cherry picked from commit 97e2028337265a5529d0b0a5fb76bb52e1a4ab37)

+++ qtfred/src/CheckedDataListModel.h
@@ -81,7 +81,7 @@ public:
 		if (parent.isValid())
 			return 0;
 
-		return items.size();
+		return static_cast<int>(items.size());
 	}
 
 	QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override {
